reowrked the ffi module to capture coverage and separation for more readability
