+++ .github/workflows/ci.yml
@@ -167,12 +167,12 @@ jobs:
         run: |
           mkdir -p coverage
           cargo llvm-cov clean --workspace
-          cargo llvm-cov nextest --workspace --no-clean \
+          cargo llvm-cov nextest --workspace --no-clean --features coverage_nightly \
             --lcov --output-path coverage/lcov.info
 
       - name: Generate HTML report (optional)
         run: |
-          cargo llvm-cov nextest --workspace --no-clean \
+          cargo llvm-cov nextest --workspace --no-clean --features coverage_nightly \
             --html --output-dir coverage/html
 
       - name: Sanity check coverage outputs
