+++ native/rust/crates/ffi/Cargo.toml
@@ -20,3 +20,6 @@ cbindgen = "0.26"
 tempfile = "3"
 httpmock = "0.7"
 serde_json = "1"
+
+[features]
+coverage_nightly = []
